requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
deta==1.1.0
dj-config-url==0.1.1
dj-rest-auth==4.0.1
Django==3.2.19
django-allauth==0.54.0
django-cors-headers==4.0.0
django-redis==5.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.4.0
